Attempting to work out workflow for Generating a DAL from an Ef Visual Designer == ERROR WHEN CREATING CONTROLLERS  ==
Attempting to figure out the EntityFramework metadata for the model and DbContext: 'Student' Unable to create an object of type 'CourseManager'. For the different patterns supported at design time, see https://go.microsoft.com/fwlink/?linkid=851728

== SOLUTION ==

Some of the EF Core Tools commands (for example, the Scaffolding commands) require a derived DbContext instance to be created at design time in order to gather details about the application's entity types and how they map to a database schema.

In most cases, it is desirable that the DbContext thereby created is configured in a similar way to how it would be configured at run time.

The tools first try to obtain the service provider by invoking

  Program.CreateHostBuilder(),
  calling Build(),
  then accessing the Services property.

Use a design-time factory, needs to be in DAL project (e.g Ex7_DAL), not in API project.

        using Microsoft.EntityFrameworkCore.Design;
        using Microsoft.EntityFrameworkCore;

        namespace Ex7_DAL
        {
            public class ApplicationDbContextFactory : IDesignTimeDbContextFactory<CourseManager>
            {
                public CourseManager CreateDbContext(string[] args)
                {
                    var optionsBuilder = new DbContextOptionsBuilder<CourseManager>();
                    optionsBuilder.UseSqlServer(CourseManager.ConnectionString);

                    return new CourseManager(optionsBuilder.Options);
                }
            }
        }

This then enables Generating controllers, which can be done in various ways depending on your needs and preferences.

## Scaffolding feature in Visual Studio:
This is the easiest way to generate controllers, especially if you are using Visual Studio.

Right-click on the Controllers folder in your project, select "Add" > "New Scaffolded Item", and choose the "API Controller with actions, using Entity Framework" template.

This will generate a fully functional CRUD controller for your model.